+++ CHANGELOG.md
@@ -17,6 +17,7 @@ v0.9.7-dev (July xx, 2021)
 * If a component to be imported has a catalog key that is not found in the internal or external catalog list then it will be skipped and logged
 * If no statements are created the resulting element/component is deleted
 * Component and System Security Plan exports pass OSCAL 1.0.0 schema validation
+* Added a proxy for parties and responsible parties for component OSCAL export
 * Coverage 6.0b1 starts to use a modern hash algorithm (sha256) when fingerprinting for high-security environments, upgrading to avoid this safety fail.
 
 * **Bug fixes**